Add files via upload
Started with two cam example from lec 12, added our files, one error left seems to be something to do with the image_transfer3.h being included in both program.cpp and global_variables.h which is also included in program.cpp. But need the image declaration in both. if we comment out the include in global_variables.h we lose that particular error but throw a bunch more due to not having the image declaration included. Im out of ideas maybe you can figure it out. I also added ShellExecute(..) and system("TASKILL...) to open and close the image_view.exe when we run the program so we dont have to do it manually. It was annoying me.